08_everything_Is_an_api/09_sql_model_with_fast_api/requirements.txt
fastapi
uvicorn
sqlmodel
bcrypt
python-jose
passlib
datetime